Organize files
Add a "src" folder with all source code files to give a hint to new
users about the purpose of those files: source code.

Add a "bin" folder with all the scripts/compiled execs to give a hint
to new users about the purpose of those files: running the source code.

Overall organize and simplify the look of the repo.